bakcend front end is 90% done for braintumor
