+++ flask-server/routes/upload.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, jsonify, request
 import os
-
+from models.braintumor_segmentation.braintumor import *
 upload_bp = Blueprint('upload', __name__)
 
 # Specify the directory where you want to save uploaded files
@@ -30,8 +30,17 @@ def add_upload():
         uploaded_file.save(file_path)
 
         # You can perform further processing on the saved file here if needed
+        image_file = request.files['file']                                          #get the image file
 
-        return jsonify({'message': 'File uploaded successfully', 'file_path': file_path})
+        image_name = os.path.basename(file_path)  #to get only the path name     
+        image_name = image_name.split('.')[0]       #name of image without file extension
+        image_location = file_path #location of the image
+        process_image(data_transforms, image_location, image_name, segment_model)
+        #calling for classification
+        #predicted_tumor = model_predict(image_location, classifier_model)
+        segmented_image_path = f"static/{image_name}-SEGMENTED.png"
+        return jsonify({'message': 'File uploaded successfully', 'file_path': file_path, 'segmented_image_path': segmented_image_path})
+ 
 
     except Exception as e:
         return jsonify({'error': f'Error uploading file: {str(e)}'}), 500
